web: limit the API of the `useCurrentSpan` hook
